move account service to application package
