+++ app/src/androidTest/java/ch/epfl/sdp/cook4me/application/AccountServiceTest.kt
@@ -1,9 +1,9 @@
-package ch.epfl.sdp.cook4me
+package ch.epfl.sdp.cook4me.application
 
 import androidx.activity.ComponentActivity
 import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import ch.epfl.sdp.cook4me.ui.AccountService
+import ch.epfl.sdp.cook4me.assertThrowsAsync
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.ktx.auth
 import com.google.firebase.ktx.Firebase
